% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{abbr}
\alias{abbr}
\title{Get the state abbreviation for a FIPS code}
\usage{
abbr(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
a \code{character} vector
}
\description{
Get the state abbreviation for a FIPS code
}
\examples{
fipio::abbr("37")
fipio::abbr("06001")

}
