% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{county}
\alias{county}
\title{Get the county name for a FIP code}
\usage{
county(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIP code}
}
\value{
a \code{character} vector
}
\description{
Get the county name for a FIP code
}
\examples{
fipio::county("37129")
fipio::county("06001")

# 2-digit FIP codes will not work
fipio::county("37")

}
