% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.intersects}
\alias{.intersects}
\title{Simple intersection via ray casting}
\usage{
.intersects(x, y, geometry)
}
\value{
indices of points in \code{x} and \code{y}
that intersect \code{geometry}
}
\description{
Simple intersection via ray casting
}
\keyword{internal}
