% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.segment_intersect}
\alias{.segment_intersect}
\title{Check if a point intersects with a side of a polygon}
\usage{
.segment_intersect(side, x, y)
}
\description{
Check if a point intersects with a side of a polygon
}
\keyword{internal}
