% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{fips_abbr}
\alias{fips_abbr}
\title{Get the state abbreviation for a FIPS code}
\usage{
fips_abbr(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
a \code{character} vector
}
\description{
Get the state abbreviation for a FIPS code
}
\examples{
fipio::fips_abbr("37")
fipio::fips_abbr("06001")

}
