% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{fips_county}
\alias{fips_county}
\title{Get the county name for a FIPS code}
\usage{
fips_county(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
a \code{character} vector
}
\description{
Get the county name for a FIPS code
}
\examples{
fipio::fips_county("37129")
fipio::fips_county("06001")

# 2-digit FIP codes will not work
fipio::fips_county("37")

}
