% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{fips_geometry}
\alias{fips_geometry}
\title{Get the geometry for a FIPS code}
\usage{
fips_geometry(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
an \code{sfg}/\code{sfc} object
}
\description{
Get the geometry for a FIPS code
}
\examples{
\dontrun{
fipio::fips_geometry("37")
fipio::fips_geometry("06001")
}

}
