% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{fips_metadata}
\alias{fips_metadata}
\title{Get the metadata for a FIPS code}
\usage{
fips_metadata(fip, geometry = FALSE)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}

\item{geometry}{If \code{TRUE}, returns a geometry column}
}
\value{
a \code{data.frame}
}
\description{
Get the metadata for a FIPS code
}
\examples{
fipio::fips_metadata("37")
fipio::fips_metadata("06001")

}
