% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{fips_state}
\alias{fips_state}
\title{Get the state name for a FIPS code}
\usage{
fips_state(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
a \code{character} vector
}
\description{
Get the state name for a FIPS code
}
\examples{
fipio::fips_state("37")
fipio::fips_state("06001")

}
