% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{fips_to_zip}
\alias{fips_to_zip}
\title{Get zip codes for a given FIPS code.}
\usage{
fips_to_zip(fip_code, ...)
}
\arguments{
\item{fip_code}{a 5-digit US FIPS code.}

\item{...}{Additional 5-digit US FIPS codes.}
}
\value{
a \code{list} with names set to the given
FIPS codes, and elements of zip codes.
}
\description{
Get zip codes for a given FIPS code.
}
\examples{
fipio::fips_to_zip("37129", "06099")
}
