% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{geometry}
\alias{geometry}
\title{Get the geometry for a FIPS code}
\usage{
geometry(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}
}
\value{
an \code{sfg}/\code{sfc} object
}
\description{
Get the geometry for a FIPS code
}
\examples{
\dontrun{
fipio::geometry("37")
fipio::geometry("06001")
}

}
