% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{metadata}
\alias{metadata}
\title{Get the metadata for a FIPS code}
\usage{
metadata(fip, geometry = FALSE)
}
\arguments{
\item{fip}{2-digit or 5-digit FIPS code}

\item{geometry}{If \code{TRUE}, returns a geometry column (requires \code{sfheaders})}
}
\value{
a \code{data.frame}
}
\description{
Get the metadata for a FIPS code
}
\examples{
fipio::metadata("37")
fipio::metadata("06001")

}
