% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipio.R
\name{state}
\alias{state}
\title{Get the state name for a FIP code}
\usage{
state(fip)
}
\arguments{
\item{fip}{2-digit or 5-digit FIP code}
}
\value{
a \code{character} vector
}
\description{
Get the state name for a FIP code
}
\examples{
fipio::state("37")
fipio::state("06001")

}
