% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{using_fastmatch}
\alias{using_fastmatch}
\title{Get the matching function that \code{fipio} is using}
\usage{
using_fastmatch()
}
\value{
\code{TRUE} if \code{fastmatch::fmatch} is used.
}
\description{
This function is primarily for debugging purposes,
or for ensuring that the correct matching function
is used.
}
