% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{zip_geometry}
\alias{zip_geometry}
\title{Get zip code geometry}
\usage{
zip_geometry(zip_code, ...)
}
\arguments{
\item{zip_code}{a 5-digit US zip code}

\item{...}{Additional zip codes}
}
\value{
an \code{sfc} object of zip_code geometries
}
\description{
Get zip code geometry
}
