% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{zip_to_fips}
\alias{zip_to_fips}
\title{Get FIPS codes for a given zip code/ZCTA}
\usage{
zip_to_fips(zip_code, ...)
}
\arguments{
\item{zip_code}{a 5-digit US zip code.}

\item{...}{Additional 5-digit US zip codes.}
}
\value{
a \code{list} with names set to the given
zip codes, and elements of FIPS codes.
}
\description{
Get FIPS codes for a given zip code/ZCTA
}
\examples{
fipio::zip_to_fips("95380", "28412")
}
